+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - (Feature) (Platform) Integration Unix Connection
 - (Feature) (Platform) Asymmetric Key Support
 - (Feature) Upgrade 4.0 process
+- (Bugfix) (Platform) Fix empty Endpoints Slice
+- (Feature) (Platform) Authn PublicKey Support
 
 ## [1.4.1](https://github.com/arangodb/kube-arangodb/tree/1.4.1) (2026-01-27)
 - (Bugfix) (Platform) Ensure Helm History limit in ArangoPlatformService